partial registrations always need to be completeable
