+++ spec/controllers/bikes_controller_spec.rb
@@ -1,6 +1,7 @@
 require 'spec_helper'
 
 describe BikesController do
+  it_behaves_like 'bikes_controller_create' # Splitting spec file up, it's too big
   describe 'index' do
     context 'no subdomain' do
       context 'legacy' do
@@ -233,36 +234,6 @@ describe BikesController do
   end
 
   describe 'new' do
-    let(:user) { FactoryGirl.create(:user) }
-    context 'legacy' do
-      before do
-        CycleType.bike
-        PropulsionType.foot_pedal
-      end
-
-      it 'renders a new stolen bike' do
-        set_current_user(user)
-        get :new, stolen: true
-        expect(response.code).to eq('200')
-        expect(assigns(:bike).stolen).to be_truthy
-      end
-
-      it 'renders a new recovered bike' do
-        set_current_user(user)
-        get :new, recovered: true
-        expect(response.code).to eq('200')
-        expect(assigns(:bike).recovered).to be_truthy
-      end
-
-      it 'renders a new organization bike' do
-        organization = FactoryGirl.create(:organization)
-        FactoryGirl.create(:membership, user: user, organization: organization)
-        set_current_user(user)
-        get :new
-        expect(response.code).to eq('200')
-      end
-    end
-
     context 'not signed in' do
       it 'sets redirect_to' do
         get :new, stolen: true, b_param_token: 'cool-token-thing'
@@ -273,12 +244,23 @@ describe BikesController do
       end
     end
 
-    context 'revised layout' do
-      before do
-        # instantiate the required bike attrs... there is a better way to do this.
-        CycleType.bike
-        PropulsionType.foot_pedal
-        set_current_user(user)
+    context 'signed in' do
+      include_context :logged_in_as_user
+      let(:manufacturer) { FactoryGirl.create(:manufacturer) }
+      let(:color) { FactoryGirl.create(:color) }
+      context 'passed stolen param' do
+        it 'renders a new stolen bike' do
+          get :new, stolen: true
+          expect(response.code).to eq('200')
+          expect(assigns(:bike).stolen).to be_truthy
+        end
+      end
+      context 'passed recovered param' do
+        it 'renders a new recovered bike' do
+          get :new, recovered: true
+          expect(response.code).to eq('200')
+          expect(assigns(:bike).recovered).to be_truthy
+        end
       end
       context 'stolen from params' do
         it 'renders a new stolen bike' do
@@ -291,17 +273,28 @@ describe BikesController do
           expect(response).to render_with_layout('application_revised')
         end
       end
-
       context 'bike through b_param' do
+        let(:bike_attrs) do
+          {
+            manufacturer_id: manufacturer.id,
+            primary_frame_color_id: color.id,
+            owner_email: 'something@stuff.com'
+          }
+        end
         context 'valid b_param' do
           it 'renders the bike from b_param' do
-            manufacturer = FactoryGirl.create(:manufacturer)
-            color = FactoryGirl.create(:color)
-            bike_attrs = {
-              manufacturer_id: manufacturer.id,
-              primary_frame_color_id: color.id,
-              owner_email: 'something@stuff.com'
-            }
+            b_param = BParam.create(params: { bike: bike_attrs.merge('revised_new' => true) })
+            expect(b_param.id_token).to be_present
+            get :new, b_param_token: b_param.id_token
+            bike = assigns(:bike)
+            expect(assigns(:b_param)).to eq b_param
+            expect(bike.is_a?(Bike)).to be_truthy
+            bike_attrs.each { |k, v| expect(bike.send(k)).to eq(v) }
+            expect(response).to render_with_layout('application_revised')
+          end
+        end
+        context 'partial registration by organization' do
+          it 'renders for the user (even though a different creator)' do
             b_param = BParam.create(params: { bike: bike_attrs.merge('revised_new' => true) })
             expect(b_param.id_token).to be_present
             get :new, b_param_token: b_param.id_token
@@ -321,327 +314,6 @@ describe BikesController do
             expect(assigns(:b_param)).to_not eq b_param
             expect(response).to render_with_layout('application_revised')
             expect(flash[:error]).to match(/sorry/i)
-          end
-        end
-      end
-    end
-  end
-
-  describe 'create' do
-    before do
-      # instantiate the required bike attrs... there is a better way to do this.
-      CycleType.bike
-      PropulsionType.foot_pedal
-    end
-    context 'legacy' do
-      describe 'web interface submission' do
-        before :each do
-          @user = FactoryGirl.create(:user)
-          @b_param = FactoryGirl.create(:b_param, creator: @user)
-          manufacturer = FactoryGirl.create(:manufacturer)
-          set_current_user(@user)
-          @bike_params = {
-            serial_number: '1234567890',
-            b_param_id_token: @b_param.id_token,
-            cycle_type_id: FactoryGirl.create(:cycle_type).id,
-            manufacturer_id: manufacturer.id,
-            rear_tire_narrow: 'true',
-            rear_wheel_size_id: FactoryGirl.create(:wheel_size).id,
-            primary_frame_color_id: FactoryGirl.create(:color).id,
-            handlebar_type_id: FactoryGirl.create(:handlebar_type).id,
-            owner_email: @user.email
-          }
-        end
-
-        it "does not use the b_param if isn't owned by user" do
-          user = FactoryGirl.create(:user)
-          set_current_user(user)
-          post :create, bike: @bike_params
-          @b_param.reload
-          expect(@b_param.created_bike_id).to_not be_present
-        end
-
-        it 'creates a new stolen bike' do
-          FactoryGirl.create(:country, iso: 'US')
-          @bike_params[:phone] = '312.379.9513'
-          expect do
-            post :create, stolen: 'true', bike: @bike_params
-          end.to change(StolenRecord, :count).by(1)
-          expect(@b_param.reload.created_bike_id).not_to be_nil
-          expect(@b_param.reload.bike_errors).to be_nil
-          expect(@user.reload.phone).to eq('3123799513')
-        end
-
-        it 'creates a new ownership and bike from an organization' do
-          organization = FactoryGirl.create(:organization)
-          FactoryGirl.create(:membership, user: @user, organization: organization)
-          @bike_params[:creation_organization_id] = organization.id
-          expect do
-            post :create, bike: @bike_params
-          end.to change(Ownership, :count).by(1)
-          expect(Bike.last.creation_organization_id).to eq(organization.id)
-        end
-      end
-
-      describe 'embeded submission' do
-        let(:organization) { FactoryGirl.create(:organization_with_auto_user) }
-        let(:user) { organization.auto_user }
-        let(:b_param) { BParam.create(creator_id: organization.auto_user_id, params: { creation_organization_id: organization.id, embeded: true }) }
-        let(:bike_params) do
-          {
-            serial_number: '69',
-            b_param_id_token: b_param.id_token,
-            creation_organization_id: organization.id,
-            embeded: true,
-            additional_registration: 'Testly secondary',
-            cycle_type_id: FactoryGirl.create(:cycle_type).id,
-            manufacturer_id: FactoryGirl.create(:manufacturer).id,
-            manufacturer_other: '',
-            primary_frame_color_id: FactoryGirl.create(:color).id,
-            handlebar_type_id: FactoryGirl.create(:handlebar_type).id,
-            owner_email: 'flow@goodtimes.com'
-          }
-        end
-        let(:testable_bike_params) { bike_params.except(:b_param_id_token, :embeded) }
-        context 'non-stolen' do
-          it 'creates a new ownership and bike from an organization' do
-            expect(user).to be_present
-            expect do
-              post :create, bike: bike_params
-            end.to change(Ownership, :count).by 1
-            bike = Bike.last
-            testable_bike_params.each do |k, v|
-              pp k unless bike.send(k).to_s == v.to_s
-              expect(bike.send(k).to_s).to eq v.to_s
-            end
-          end
-        end
-        context 'stolen' do
-          let(:state) { FactoryGirl.create(:state) }
-          let(:country) { state.country }
-          let(:stolen_params) do
-            {
-              country_id: country.id,
-              street: '2459 W Division St',
-              city: 'Chicago',
-              zipcode: '60622',
-              state_id: state.id,
-              date_stolen_input: Date.today.strftime('%m-%d-%Y')
-            }
-          end
-          context 'valid' do
-            it 'creates a new ownership and bike from an organization' do
-              expect do
-                post :create, bike: bike_params.merge(stolen: true), stolen_record: stolen_params
-              end.to change(Ownership, :count).by 1
-              bike = Bike.last
-              testable_bike_params.each { |k, v| expect(bike.send(k).to_s).to eq v.to_s }
-              stolen_record = bike.current_stolen_record
-              stolen_params.except(:date_stolen_input).each { |k, v| expect(stolen_record.send(k).to_s).to eq v.to_s }
-              expect(stolen_record.date_stolen.to_date).to eq Date.today
-            end
-          end
-          context 'invalid' do
-            it 'renders the stolen form with all the attributes' do
-              target_path = embed_organization_path(id: organization.slug, b_param_id_token: b_param.id_token)
-              expect do
-                post :create, bike: bike_params.merge(stolen: '1', primary_frame_color: nil),
-                              stolen_record: stolen_params
-              end.to change(Ownership, :count).by 0
-              expect(response).to redirect_to target_path
-              bike = assigns(:bike)
-              testable_bike_params.except(:primary_frame_color_id).each { |k, v| expect(bike.send(k).to_s).to eq v.to_s }
-              expect(bike.stolen).to be_truthy
-              # we retain the stolen record attrs, it would be great to test that they are
-              # assigned correctly, but I don't know how - it needs to completely
-              # render the new action
-            end
-          end
-        end
-      end
-
-      describe 'extended embeded submission' do
-        it 'registers a bike and uploads an image' do
-          Sidekiq::Testing.inline! do
-            organization = FactoryGirl.create(:organization)
-            user = FactoryGirl.create(:user)
-            FactoryGirl.create(:membership, user: user, organization: organization)
-            organization.save
-            manufacturer = FactoryGirl.create(:manufacturer)
-            b_param = BParam.create(creator_id: organization.auto_user.id, params: { creation_organization_id: organization.id, embeded: true })
-            test_photo = Rack::Test::UploadedFile.new(File.open(File.join(Rails.root, 'spec', 'fixtures', 'bike.jpg')))
-            expect_any_instance_of(ImageAssociatorWorker).to receive(:perform).and_return(true)
-            bike_params = {
-              serial_number: '69',
-              b_param_id_token: b_param.id_token,
-              creation_organization_id: organization.slug,
-              embeded: true,
-              embeded_extended: true,
-              cycle_type_id: FactoryGirl.create(:cycle_type).id,
-              manufacturer_id: manufacturer.name,
-              primary_frame_color_id: FactoryGirl.create(:color).id,
-              handlebar_type_id: FactoryGirl.create(:handlebar_type).id,
-              owner_email: 'Flow@goodtimes.com',
-              image: test_photo
-            }
-            post :create, bike: bike_params
-            expect(response).to redirect_to(embed_extended_organization_url(organization))
-            expect(Bike.last.owner_email).to eq bike_params[:owner_email].downcase
-          end
-        end
-      end
-
-      describe 'extended embed submission with persisted email' do
-        it 'registers a bike and redirects with persist_email' do
-          organization = FactoryGirl.create(:organization)
-          user = FactoryGirl.create(:user)
-          FactoryGirl.create(:membership, user: user, organization: organization)
-          organization.save
-          manufacturer = FactoryGirl.create(:manufacturer)
-          b_param = BParam.create(creator_id: organization.auto_user.id, params: { creation_organization_id: organization.id, embeded: true })
-          bike_params = {
-            serial_number: '69',
-            b_param_id_token: b_param.id_token,
-            creation_organization_id: organization.id,
-            embeded: true,
-            embeded_extended: true,
-            cycle_type_id: FactoryGirl.create(:cycle_type).id,
-            manufacturer_id: manufacturer.id,
-            primary_frame_color_id: FactoryGirl.create(:color).id,
-            handlebar_type_id: FactoryGirl.create(:handlebar_type).id,
-            owner_email: 'Flow@goodtimes.com'
-          }
-          post :create, bike: bike_params, persist_email: true
-          expect(response).to redirect_to(embed_extended_organization_url(organization, email: 'flow@goodtimes.com'))
-        end
-      end
-    end
-
-    context 'revised' do
-      let(:user) { FactoryGirl.create(:user) }
-      let(:manufacturer) { FactoryGirl.create(:manufacturer) }
-      let(:color) { FactoryGirl.create(:color) }
-      before do
-        set_current_user(user)
-      end
-      context 'no existing b_param and stolen' do
-        let(:wheel_size) { FactoryGirl.create(:wheel_size) }
-        let(:country) { Country.united_states }
-        let(:state) { FactoryGirl.create(:state, country: country) }
-        let(:bike_params) do
-          {
-            b_param_id_token: '',
-            cycle_type_id: CycleType.bike.id.to_s,
-            serial_number: 'example serial',
-            manufacturer_other: '',
-            year: '2016',
-            frame_model: 'Cool frame model',
-            primary_frame_color_id: color.id.to_s,
-            secondary_frame_color_id: '',
-            tertiary_frame_color_id: '',
-            owner_email: 'something@stuff.com',
-            phone: '312.379.9513',
-            stolen: true
-          }
-        end
-        let(:stolen_params) do
-          {
-            country_id: country.id,
-            street: '2459 W Division St',
-            city: 'Chicago',
-            zipcode: '60622',
-            state_id: state.id
-          }
-        end
-        before do
-          expect(BParam.all.count).to eq 0
-        end
-        context 'successful creation' do
-          it "creates a bike and doesn't create a b_param" do
-            success_params = bike_params.merge(manufacturer_id: manufacturer.slug)
-            bb_data = { bike: { rear_wheel_bsd: wheel_size.iso_bsd.to_s }, components: [] }.as_json
-            # We need to call clean_params on the BParam after bikebook update, so that
-            # the foreign keys are assigned correctly. This is how we test that we're
-            # This is also where we're testing bikebook assignment
-            expect_any_instance_of(BikeBookIntegration).to receive(:get_model) { bb_data }
-            expect do
-              post :create, stolen: true, bike: success_params.as_json, stolen_record: stolen_params
-            end.to change(Bike, :count).by(1)
-            expect(flash[:success]).to be_present
-            expect(BParam.all.count).to eq 0
-            bike = Bike.last
-            bike_params.delete(:manufacturer_id)
-            bike_params.delete(:phone)
-            bike_params.each { |k, v| expect(bike.send(k).to_s).to eq v.to_s }
-            expect(bike.manufacturer).to eq manufacturer
-            expect(bike.stolen).to be_truthy
-            user.reload
-            expect(user.phone).to eq '3123799513'
-            expect(bike.current_stolen_record.phone).to eq '3123799513'
-            stolen_record = bike.current_stolen_record
-            stolen_params.delete(:state_id) # this doesn't show up, don't care for now, shows up for real
-            stolen_params.each { |k, v| expect(stolen_record.send(k).to_s).to eq v.to_s }
-          end
-        end
-        context 'failure' do
-          it 'assigns a bike and a stolen record with the attrs passed' do
-            expect do
-              post :create, stolen: true, bike: bike_params.as_json, stolen_record: stolen_params
-            end.to change(Bike, :count).by(0)
-            expect(BParam.all.count).to eq 1
-            bike = assigns(:bike)
-            bike_params.delete(:manufacturer_id)
-            bike_params.delete(:phone)
-            bike_params.each { |k, v| expect(bike.send(k).to_s).to eq v.to_s }
-            expect(bike.stolen).to be_truthy
-            # we retain the stolen record attrs, it would be great to test that they are
-            # assigned correctly, but I don't know how - it needs to completely
-            # render the new action
-            # stolen_record = assigns(:stolen_record)
-            # stolen_params.each { |k, v| expect(stolen_record.send(k).to_s).to eq v.to_s }
-          end
-        end
-      end
-      context 'existing b_param' do
-        context 'no bike' do
-          it 'creates a bike' do
-            bike_params = {
-              cycle_type_id: CycleType.bike.id.to_s,
-              serial_number: 'example serial',
-              manufacturer_other: '',
-              year: '2016',
-              frame_model: 'Cool frame model',
-              primary_frame_color_id: color.id.to_s,
-              secondary_frame_color_id: '',
-              tertiary_frame_color_id: '',
-              owner_email: 'something@stuff.com'
-            }.as_json
-            b_param = BParam.create(params: { 'bike' => bike_params })
-            bb_data = { bike: {} }
-            # We need to call clean_params on the BParam after bikebook update, so that
-            # the foreign keys are assigned correctly. This is how we test that we're
-            # This is also where we're testing bikebook assignment
-            expect_any_instance_of(BikeBookIntegration).to receive(:get_model) { bb_data }
-            expect do
-              post :create, bike: { manufacturer_id: manufacturer.slug, b_param_id_token: b_param.id_token }
-            end.to change(Bike, :count).by(1)
-            expect(flash[:success]).to be_present
-            bike = Bike.last
-            expect(bike.creator_id).to eq user.id
-            b_param.reload
-            expect(b_param.created_bike_id).to eq bike.id
-            bike_params.delete(:manufacturer_id)
-            bike_params.each { |k, v| expect(bike.send(k).to_s).to eq v }
-            expect(bike.manufacturer).to eq manufacturer
-          end
-        end
-        context 'created bike' do
-          it 'redirects to the bike' do
-            bike = FactoryGirl.create(:bike)
-            b_param = BParam.create(params: { bike: {} }, created_bike_id: bike.id, creator_id: user.id)
-            expect(b_param.created_bike).to be_present
-            post :create, bike: { b_param_id_token: b_param.id_token }
-            expect(response).to redirect_to(edit_bike_url(bike.id))
           end
         end
       end
